pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "librosa>=0.11.0",
    "numpy>=2.2.6",
    "paddlepaddle>=3.0.0",
    "paddlespeech>=1.5.0",
    "pyttsx3>=2.98",
    "soundfile>=0.13.1",
]
